#9 - Alterando o tipo de retorno do método executar

+++ src/main/java/com/imd/config/mudanca/banco/command/Operacao.java
@@ -6,5 +6,5 @@ import com.imd.config.mudanca.banco.domain.Conta;
 
 public interface Operacao {
 
-		public double executar(Conta contaOrigem, Conta contaDestino, BigDecimal valor);
+		public BigDecimal executar(Conta contaOrigem, Conta contaDestino, BigDecimal valor);
 }
